Fix showing labels of newly added wxNotebook pages in wxOSX
The existing code added new tabs before setting their label, which
resulted in them being shown initially truncated.

Fix this by adding new tabs with their correct label (and image), as
this seems to make more sense and also fixes the problem.

Also don't override SetMaximum() in wxCocoaTabView, this doesn't make
sense as wxNotebook doesn't have a "maximum" property unlike wxSpinCtrl
or wxGauge, for which this function (and the associated SetMinimum())
exists, and just do what this function did in SetupTabs() itself, which
is the function explicitly intended to be overridden in wxNotebook.

Closes #25194.