Avoid using wxWindow::SetLabel() in wxOSX wxNotebook code
This function doesn't work for all windows, notably it doesn't do
anything and triggers an assert for wxTextCtrl since the changes of
79f06970a8 (Make wxTextCtrl::SetLabel() consistently do nothing
everywhere, 2024-11-28), so don't use it to store the notebook page
title and store it separately, just as we already do for the page
images and as was done for wxGTK in the recent 3835619d76 (Store the
original page text in wxGTK wxNotebook, 2025-02-24).

Incidentally, this also makes GetPageText() return the same string that
was passed to SetPageText() instead of returning the string with the
mnemonics stripped, as it did before.

Closes #25195.