Remove buggy and mostly unused code for wxEVT_KEY_UP in wxGTK
Assuming that we always get a GDK_KEY_RELEASE event after a matching
GDK_KEY_PRESS one, as was done in this code added back in 654d905270
(attempts to process non ASCII key strokes properly, 2001-10-10) was
wrong, as this isn't the case when pressing multiple keys and not
releasing them in exactly the LIFO order (e.g. pressing "Shift", then
"5" and then releasing "Shift" before "5").

We could generalize this by using a stack, but this code shouldn't be
necessary in the first place, as we really shouldn't be using
GdkEventKey::string anyhow.