Ignore unknown events read from IOCP in wxMSW wxFileSystemWatcher
We can get events with unknown (and, arguably, invalid) values, at least
when using Samba, so don't assert if this happens but just ignore them.

This extends the changes of 83c75a439d (Ignore invalid file notification
events, 2020-10-30) to cover all unknown events and not just those with
"Action" field set to 0.

Closes #18953.