Do generate CHAR events for Ctrl+[,\,],^ and _ in wxGTK
Also simplify the code by getting rid of AdjustCharEventKeyCodes()
because we don't need it at all when handling IM input: in this case we
know that we have a valid Unicode characters and that Control modifier
is not used.

Instead, determine the key and character codes to use directly in the
code generating wxEVT_CHAR events.