Forbid dragging tabs in invalid position in another tab control
We already checked that dragging tabs inside the same tab control was
only allowed when they remained among the tabs of the same kind (e.g.
pinned or normal), but the same check was missing when dragging to a
different tab control, allowing to drag a normal tab into the group of,
or even before, the pinned tabs.

Check for this too now to prevent this from happening as well.