Fix wxSearchCtrl best size computation in wxGTK
The size returned by the default implementation was too small, at least
partially due to the fact that GtkEntry doesn't seem to account for the
icons shown by GtkSearchEntry.

Ensure that we have at least some minimum width to make it possible to
enter text into this control, even when it's created with the default
size.

Closes #25210.