Respect best control size when adding it to wxGTK wxToolBar
Usually, the best size is used by wx own layout code, but it isn't used
when a control is part of GtkToolbar and in this case the widget used
its default GTK size, ignoring any calls to SetMinSize().

Fix this by ensuring that it gets at least its best size.

See #25210.