Fix crash in wxTreeListCtrl when using wxGTK wxDataViewCtrl
Since the changes of 0f607ed3f6 (Simplify FindNode() and reuse it in
FindParentNode(), 2024-12-23) wxTreeListModel::GetParent() could be
called with an invalid item, resulting in a crash there.

Add back the check for the item being valid to wxDataViewCtrl code to
avoid this, even if wxTreeListModel function shouldn't crash when given
this input, it seems that there could be other models with the same
problem, so it's safer to avoid calling them in this case.

Co-Authored-By: Daniel Kulp <dan@kulp.com>