Use --undefined-version with --version-script if supported
Currently we have a single version-script file which is used when
linking all libraries, meaning that we'll always have some versions
defined for the symbols not present in the library being linked.

With --undefined-version, which has been the default until recently,
versions defined for such symbols were just ignored, but since
--no-undefined-version became default, they started resulting in link
errors.

The proper fix would be to define per-library version scripts, but for
now just avoid the error by using --undefined-version explicitly if it
is supported (note that this is a relatively recent addition, as it only
appeared in GNU bintools 2.40 released in 2022, so we can't assume it
always is supported).