Fix dangling references to temporaries in internat sample
In some build variants (e.g. when wxUSE_INTL==0) _() is a macro doing
nothing, in particular not involving a function call, and the temporary
wxString created by it doesn't survive past the end of the statement and
hence a reference to it can't be used.

Just make a copy of the string to fix this.