Restore correct modifiers order in menu accelerators under MSW
The changes of 4dcda62f9d (Normalize displayed accelerators in wxMSW
too, 2022-12-04) accidentally switched the order in which modifiers were
displayed in menu accelerators in wxMSW.

Fix this by reordering code in wxAcceleratorEntry to return the
accelerator text with the modifiers in MSW native order. Note that other
platforms don't use the same order, but they don't use this function
either, so this change doesn't affect them.