Fix using C++ objects from block in wxWebRequestURLSession code
The code added in cd05d0a788 (Partially implement sync requests support
in NSURLSession backend, 2024-08-16) was wrong as it copied Semaphore
object into the block storage, resulting in destroying the dispatch
semaphore twice.

Fix this by making Semaphore non-copyable, as it should have always
been, and using a unique_ptr<> to access it.

Also implement copying properly for TaskResult, which is also used both
in the block and outside of it.

Closes #25231.