Add wxAuiCreateBitmap() overload taking SVG data
Use it to define dock art icons using SVG, so that they look nicer in
high DPI.

Note that the same "close" icon is used on all platforms when using SVG,
unlike with the previously XBM icons, as none of the icons really
corresponds to the native platform style, so it doesn't seem to be worth
it to make this platform-dependent.

The best thing to do to allow applications to customize them would be to
add these icons to wxArtProvider, but leave this for the future.