Handle deserializing wxAuiNotebook data without all pages
This allows newer versions of the program, adding new pages, to still
restore data saved by the older versions.

By default the pages not described by the deserialized data are simply
appended to the main tab control, however this behaviour can be
customized by overriding the new HandleOrphanedPage() function.

The signature of the new function is rather complex and not very
elegant, but this seems to be the least bad solution, as it allows wxAUI
itself to provide the default values (main tab control and -1, meaning
to append, for the index), making the code of the applications
overriding this function simpler.

Also note that we really need to get the location for the page in some
way, e.g. we couldn't just let HandleOrphanedPage() or a similar
function put it into its place on its own because at the moment of its
call the wxAuiNotebook invariant saying that each page in m_tabs is
present in exactly one wxAuiTabCtrl is broken, and so many wxAuiNotebook
functions, including e.g. DeletePage(), don't work as expected.