Add static wxApp::CallOnUnhandledException() wrapper and use it
The new function simply checks if wxTheApp exists and calls
OnUnhandledException() on it if it does. It also handles any exceptions
thrown by the overridden OnUnhandledException(), which was previously
done in a single place only but should be really done everywhere (or,
alternatively, nowhere).

Use it instead of calling OnUnhandledException() directly: an extra
checks for wxTheApp should do no harm and may prevent a crash in some
rare scenarios.