Fix notebook art EVT_UPDATE_UI handler in the AUI sample
The existing code was confusing because it did something strange for
ID_NotebookArt{Gloss,Simple} in OnUpdateUI(), but this handler wasn't
actually connected to the menu items with these IDs.

Do connect it and add another member variable storing the current choice
of the art provider, as the existing handler code couldn't, and didn't,
work (and this only wasn't a problem because it was never executed).