wxQt: Fix handling command line arguments in QApplication and wx
For example: "./minimal -style fusion" or "./minimal -stylesheet myStylesheet"
didn't work before, but it work now with this commit.

Explanation:
------------
The old code was not correct: it just crashes when trying to delete "argv_"
and if it doesn't, it tries to recreate it with the already consumed args
which doesn't make sense to wxWidgets.

Additionally, it makes unnecessary memory allocations for the arguments,
which simply results in memory leaks.

Closes #25269.