Move ShowDropDown() implementation to wxAuiTabArtBase too
This function was implemented slightly differently in the existing tab
art subclasses, with wxAuiSimpleTabArt using check menu items and
positioning the menu differently, but also not accounting for the
ampersands in the page caption.

None of these differences looked intentional, and having many check
items in the menu didn't look good, so reuse wxAuiGenericTabArt code for
both classes.

This commit is best viewed with Git --color-moved option.