Share DrawButton() implementation between tab art classes too
The existing implementations were slightly different, but the difference
can be extracted into a new DrawButtonBitmap() virtual function, which
is trivial by default and whose reimplementation in wxAuiSimpleTabArt
already existed as a global DrawButtons() function in the source, so it
just had to be renamed and moved.

The new version duplicates less code and will allow reusing DrawButton()
in other tab art classes.