Fix printing selected pages when overriding GetPageInfo()
Simply ignore the "from" and "to" values returned from this legacy
function as wxMSW always did it and nothing else makes much sense, as we
don't want to override the page range(s) entered by the user in the
print dialog.

Update the documentation to clarify this unexpected behaviour and also
provide a better explanation of what the new GetPagesInfo() should, and
should not, do.

Closes #25254 (again).