Avoid giving an error after cancelling file dialog when printing
Cancelling the dialog shown to select the file to print to, when
printing to a file, shouldn't result in an error message, so fix this by
checking last error after calling Win32 StartDoc() which does it.

As there is no clean way to pass this error from wxPrinterDC::StartDoc()
back to wxWindowsPrinter, which calls it indirectly via wxPrintout, just
pass it via a global variable -- this is ugly, but should work fine as
we can't be printing more than one thing at the same time.