Avoid bogus SetFocus() debug errors while showing print dialog
Getting WM_ACTIVATE just before the print dialog dismissal but while it
still existed resulted in spurious errors about not being able to
restore the last focus in the previously active window because it didn't
recover its active status yet, at this moment.

Use an ugly hack to ignore activation events happening in this case, as
there doesn't seem to be anything else that we can do to avoid getting
them.

Closes #13089.