Generate key codes in US layout for OEM keys in wxMSW
Instead of not generating any key code at all for such keys, generate
the key code that the same key would produce in the US layout.

This is especially useful for the keyboard using Cyrillic layout, as
they would typically have both sets of characters on them and it makes
sense to generate an event with the key code of e.g. `;` for the key
with this marking even if it produces a Cyrillic letter when pressed.

Update the documentation to document this behaviour, which was already
implemented in wxGTK, and generally try to explain things more clearly.

Note that this doesn't affect accelerators, which are handled by
Windows, i.e. this isn't sufficient to fix #22058 under MSW.