Allow triggering more accelerators in non-US keyboard layouts
If the character used in the accelerator (e.g. "]" used in "Ctrl-]") is
not present in the layout, use the key that would have generated it in
the US layout instead.

This is arguably better than nothing, but it still doesn't allow to
generate all the possible key combinations, so update the documentation
to warn about it.