Use wxMSW wxFileDialog implementation in wxQt under MSW
This adds support for custom controls in the file dialog, unavailable in
the Qt version.

Note that this requires also using wxMSW implementation of wxDirDialog
because in both wxMSW and wxQt wxFileDialog is tightly intertwined with
it, even though it doesn't have any known advantages compared to the Qt
version (but OTOH it doesn't have any known disadvantages either).