Fix wxTreeCtrl indent regression
The changes from #24640 don't seem to work when the size of the images in the image list is too small.
The win32 control seems to expect an image size of at least 16px, corrected for the DPI.
When the images are smaller, for example just 16px, the extra indent causes the expand/collapse hitbox
to be shifted to the right and not overlap with the twist (or plus-min) button.

Fix this by reducing the indent if the images are too small.