Replace wxPrintDialogData booleans with a single enum
It was unnatural to have 3 booleans encoding 4 different states, as this
made a lot of impossible states representable.

Change this by using a single 4-state enum instead.

The price to pay is that SetAllPages(false) etc now behave somewhat
strangely: they don't do anything when not printing all pages (same as
before) and set the print target to "only specified pages" otherwise.
This is almost but not quite the same as the old behaviour because
calling

        SetAllPages(true);
        SetSelection(true);
        SetAllPages(false);

resulted in GetSelection() being true before but GetSpecifiedPages()
being true now. But hopefully not many people did anything like this and
at least this makes the state after the first 2 calls well-defined,
unlike before.

Also make printing all pages default everywhere and not just under Mac
because nothing else seems to make sense.

Update the documentation and, in particular, add the documentation of
SetAllPages() which somehow was never documented before.