Remove obsolete assert about menuitem ID zero on Mac
The assert, introduced in 408aef5 (as we cannot handle zero commands
properly, we have to warn, 2005-01-04) says that menuitem ID zero
does not work on Mac, yet empirically it seems to work just fine.

If it really doesn't work, then this assert should be present in
all other ports too, to help writing portable code.