Define ssize_t under Win32 in the same way as the platform SDK
For compatibility with the other libraries, define ssize_t in the same
way as SSIZE_T is defined in the platform headers: this was also true
for Win64, but for Win32 we defined ssize_t as int while everybody else
defined it as long, so do it too now.

See #25313.

Closes #25309.