Add wxSsize() and use it in the code
This trivial function just performs the cast of size() of some container
to int.

Using it makes it possible to write the code using int indices without
(or with fewer) casts.

It doesn't check that the cast doesn't truncate currently, although
perhaps it should -- but then it seems very unlikely that any of the
containers we work with in wx can ever get close to INT_MAX size.