Fix ABI breakage for versioned symbols in 3.2.7
Add 3.2 versions of the symbols whose version was "fixed" in ed7defadd2
(Fix multiple errors in the version script, 2025-01-01), breaking ABI as
they were not available in their old version any more.

Add wxELF_VERSION() macro which allows to define the symbol with the
given ELF version and use it, in conjunction with the existing version
in the version script, to have both WXU_3.2.N and WXU_3.2 versions of
all the symbols affected by the changes in that commit.

Closes #25327.