Honour certificate-related environment variables used by curl
Use the same environment variables that curl uses to customize the
location of the trusted certificates when using libcurl.

This is useful when using libcurl under Windows (where there is no
reasonable default) but also under Linux where a custom-built libcurl
may refer to a file or directory not available on the target system.