Do not activate CLOSE button on Cmd-C in wxOSX
Extend the hack of 22bcdf020e (Don't use Cmd+C accelerator in dilogs on
Mac by default., 2016-07-17) (later amended by 7a45b7948a (Move the hack
for Cmd+C in wxOSX to a better place, 2019-01-09), to wxID_CLOSE in
addition to wxID_CANCEL and for the same reason: Cmd-C is used for
copying and shouldn't active the button.

See #15678.

Closes #25346.