Move check for CSD for non-resizeable window in GTKDoGetSize()
For whatever reason, we don't account for the CSD for non-resizeable
windows, so simplify things by checking for this in GTKDoGetSize()
itself instead of having to use HasClientDecor() before calling it.

In one case we called GTKDoGetSize() unconditionally, but this seems to
have been a mistake as there doesn't seem to be any reason to handle
them differently just there.