Fix handling window total size under GNOME with X11
Due to getting two notifications about _NET_REQUEST_FRAME_EXTENTS from
mutter (GNOME WM), we didn't do the right thing as the second one
arrived after mapping the window and the code assumed that in this case
the window already has the correct size and only internal size needs to
be adjusted.

In this case, however, the window doesn't have the correct size and we
need to adjust it once we get the real decoration sizes values and not
just zeroes for all of them that we get during the first notification.

Note that we can't just filter out that first notification because then
we never map the window and so never get the second one.

Closes #25348.