Ensure that we include built-in 3rd party libraries headers
Use template_append when using 3rd party library template to ensure that
the include directories defined in the <lib> using this template appear
first in the compiler command line, so that the correct headers get
included even when the system libraries are also available.

Note that this required refactoring the old 3rdparty_lib template into
the base 3rdparty_lib_base and new 3rdparty_lib_u, as appending the old
template to wxregex lib overwrote libname defined in the latter
otherwise.

Closes #25350.

See #25352.