Simplify wxWindowGTK::GTKUpdateCursor()
This function had a very complicated signature and was used in
completely unreadable way, as nothing indicated what a call to
GTKUpdateCursor(false, true) did and how it differed from a call to
GTKUpdateCursor(true, false).

Split it into 3 different functions, corresponding to the different use
cases and replace the only remaining bool argument remaining in this
function with a GdkCursor pointer: while it's only used to check if it's
null or not (i.e. is effectively a boolean), passing a valid pointer or
null makes the code using this function much more understandable.

There should be no changes in behaviour in this commit, it only removes
tautologically true (or false) boolean conditions uncovered by this
refactoring.