Cache the last returned wxCursor in wxCursorBundle
This is required in order to use the cursor returned by GetCursorFor()
with wxEVT_SET_CURSOR in wxMSW: without it, the cursor object itself is
destroyed too early and HCURSOR we return to the system from this event
handler becomes invalid.

It is also more efficient, which is important for a function which may
be called repeatedly on each mouse move.