Deprecate wxBusyCursorSuspender
This class only ever worked in wxGTK and wxX11 and doesn't seem really
useful: it's purported goal is to prevent busy cursor from being shown
for the modal dialogs, but nothing special needs to be done for this as
modal dialogs created after calling wxBeginBusyCursor() don't inherit
the busy cursor and so don't show it anyhow.

Just stop doing anything at all in this class and also in the functions
GetStoredCursor() and GetBusyCursor() that were only used by it.