Restore compatibility with overriding wxApp::OnAssert()
The changes of 3ec4a23f50 (show the function in which the assert failure
occured if the compiler supports it, 2006-03-21) left the old OnAssert()
function for compatibility, but it was never called since then as the
code only called the new function.

Fix this only 19 years later by calling the old function and using a
global variable to help it pass the function name to the new one by
default.

Note that GitHub code search shows that there are a few different
projects overriding OnAssert() in their wxApp-derived class, so this fix
is not as useless as could be believed.