Do generate wxEVT_CHAR for Ctrl-Letter even in non US layouts
Use the key translated to the US layout by XKB and not the actual
character produced by the key in the current layout when determining
whether to generate wxEVT_CHAR for key presses with Ctrl or not.

This ensures that we produce the expected events for Ctrl-Latter in all
layouts and not just the ones using Latin letters.

Closes #25384.