Take bitmap scale factor into account when creating wxCursor
We need to use gdk_cursor_new_from_surface() for creating the cursor in
order to be able to specify the bitmap scale factor (note that because
gdk_cursor_new_from_pixbuf() is implemented in terms of this function
anyhow, we don't gain anything from not calling it even if the scale
factor is 1).

This fixes unwanted cursor scaling by GTK/Wayland which made the cursor
both blurry and of a wrong size.