Fix cursor bitmap size and scaling factor
Always use bitmap with the same scaling factor as the currently used one
but select the bitmap of the size corresponding to the scale of the
cursor (and not the default size of this cursor bundle itself).

This ensures that our cursor bitmaps are not rescaled and that we still
use a bigger bitmap even in standard DPI if the user increased the
default cursor size.