Prefer larger cursor in case of a tie between available sizes
Unlike for the bitmaps, using a larger cursor seems preferable to using
a smaller one in case of a tie (e.g. when using 150% DPI scaling with
only standard and @2x versions of the bitmap being available) and this
behaviour is consistent with what MSW does for the default cursor.