Try using -Wsuggest-override in the CI builds
This is a useful warning and we want to ensure that wx compiles cleanly
with it.

Note that it can't be used for gcc < 11 (it is supported by it since 5.1
but there are tons of false positives for the macros from wx/rtti.h with
the older versions) and only relatively recent clang supports this
gcc-compatible option.