Fix test for .symver asm directive support in configure
We need to use it with the correct, i.e. existing, symbol name, which
means the mangled name actually used by the linker and not just the name
of the function as it appears in the source.

Fix the test to actually detect lack of support for .symver when using
clang with -flto.

See #25438.

(cherry picked from commit f2b844307d84d34a65c8e527ff9b94bc0d7972bb)