Add wxGLCanvas::SetSwapInterval() and GetSwapInterval()
Allow setting the swap interval value or not setting it at all instead
of always forcing it to 0.

The changes of 3dde6bdeb0 (Disable swap interval in GLX wxGLCanvas
implementation too, 2023-12-25) are not always desirable, so allow to
customize the behaviour.

Implement this for Unix (GLX and EGL) backends for now and allow testing
using these functions in the cube sample.

See #24163, #24165.