Fix wxGLCanvasEGL buffer scale desyncing from the GTK scale factor
Previously, wxGLCanvasEGL only set the buffer scale of the wl_surface once during the creation of the canvas.
However, the GTK widget scale factor may change at any time (for example when the window is moved across monitors), which results in the EGL window size becoming out-of-sync with the surface's buffer scale when the canvas is resized.
This patch fixes this issue by updating the wl_surface buffer size when the canvas widget is resized.