Fix wxGLCanvasEGL scale under Wayland in high DPI
Ensure that the buffer scale is always the same as the GTK scale factor:
Previously, wxGLCanvasEGL only set the buffer scale of the wl_surface
once during the creation of the canvas. However, the GTK widget scale
factor may change at any time (for example when the window is moved
across monitors), which results in the EGL window size becoming
out-of-sync with the surface's buffer scale when the canvas is resized.

Fix this issue by updating the wl_surface buffer size whenever the canvas
widget is resized.

Closes #23733.

Closes #25465.