Disable wxUSE_FSVOLUME in the ports not supporting it
It doesn't make sense to have wxUSE_FSVOLUME=1 in wxGTK and other Unix
ports that don't provide this class, so leave it as 0 there.

Also update the documentation to say that wxFSVolume is not implemented
under Unix, not that it doesn't make sense there because it could be
mapped to file system mount points, in theory.