Define USE_WIN32=1 earlier in configure
This allows to use it to guard Windows-specific options.

Also drop the test for "doesnt_want_win32" which, confusingly, was not
the negation of the test for "wants_win32" and required a special
comment to explain this: instead, just don't define "wants_win32" and
reset USE_WIN32 in this case.