Add wxNODISCARD to stc.h.in and not only to the generated file
This wxNODISCARD was added to stc.h only in 5438f42b4a (Add wxNODISCARD
to all Clone*() functions that return a pointer, 2025-04-29) and so
would have been lost when this file is regenerated by gen_iface.py.

Add wxNODISCARD to the manually maintained file too to ensure that this
doesn't happen.

See #25354.