Fix fatal assert when using box indicators in wxStyledTextCtrl
This assert, generated from wxEVT_PAINT handler, happened because
SurfaceImpl class unconditionally casts its wxDC to wxMemoryDC before
creating wxGraphicsContext from it, but the actual wxDC was a wxPaintDC,
created in wxStyledTextCtrl::OnPaint(), and not a wxMemoryDC.

Fix this by creating wxBufferedPaintDC, which is a wxMemoryDC, instead.

Note that it is not clear how could this work before as OnPaint()
created wxPaintDC since always, and SurfaceImpl seems to have always
needed wxMemoryDC too. It would also be better to avoid casting wxDC to
wxMemoryDC in the first place, as this forces us to use wxMemoryDC even
on platforms using double-buffering, i.e. effectively unnecessarily
triple-buffering drawing here.