Don't inherit from wxNavigationEnabled<> in wxVListBox
Keyboard navigation should not be used in this simple control-like
class, but was because it derived from wxNavigationEnabled, via the
parent wxVScrolledWindow which, contrary to its name, derived from
wxPanel (and hence wxNavigationEnabled<>) rather than wxWindow.

Refactor wxVScrolledWindow to make it symmetric with wxScrolledWindow
i.e. make it a specialization of a more general wxVScrolled<> template
class and also defined wxVScrolledCanvas, for consistency with
wxScrolledCanvas, which is wxVScrolled<wxWindow>, and make wxVListBox
derive from it instead.

Closes #25588.