Avoid repainting wxNotebook when it is being destroyed
This is unnecessary and may result in a crash if a WM_PAINT event comes
after wxNotebook itself has been already destroyed (e.g. from the base
class dtor when it destroys the notebook children which triggers
invalidating the parent) because our OnPaint() depends on the object
being in a well-defined state.

Closes #25499.