Don't bother resetting wxStaticBitmap image when destroying it
This results in sending a WM_PAINT to the control parent which may be
rather unexpected if it happens while it is itself being destroyed, see
the fix for #25499 in the parent commit, so don't do it when Free() is
called from dtor.

And it is also unnecessary to do it when Free() is called from
DoUpdateImage(), as it calls MSWReplaceImageHandle() with another,
valid, handle right next anyhow.