Drop support for using generic wxHyperlinkCtrl in wxMSW
We require the use of comctl32.dll v6 now, so we can always use the
native version, remove the code falling back to the generic one.

Note that wxHyperlinkCtrl still inherits from wxGenericHyperlinkCtrl
because it uses many of its member variables and its support for popup
menu, which is not available in the native control.