Fix wxHyperlinkCtrl::GetLabel() in wxMSW native version
We wrongly returned the full HTML fragment used by the native control as
the label instead of returning just the actual label part after creating
the control.

This fixes the problem which could be easily seen by switching from the
native to the generic control in the widgets sample: the label changed
to HTML markup instead of staying the same.