Use foreground window colour as link colour in wxMSW
Make wxHyperlinkCtrl behaviour in wxMSW the same as in the generic
version and change the (normal) link colour if the foregound window
colour is changed.

Note that the control still doesn't inherit the default foreground
colour from its parent and keeps using the default link colour by
default. But it changes it if SetForegroundColour() is explicitly
called.