Remove wxAcceleratorRefData::m_ok
Don't allocate wxAcceleratorRefData at all if we don't have a valid
HACCEL for it: this is simpler and less error-prone than having a
separate m_ok member and remembering to set it to true only if m_accel
is valid (which was forgotten in at least SetHACCEL()).

There should be no observable changes.