Search menu items only in menus belonging to wx
We can't assume that all menus contain wxMenuItems, as this is not the
case for at least the per-window menu in the MDI applications and
interpreting its items user data as wxMenuItem pointers resulted in a
crash, which not searching in this menu fixes.

Having the right menu to search in also avoids the need for recursion,
which simplifies the code even further.

Closes #25522.