Simplify statbar sample by defining the frame base class once
Avoid multiple USE_MDI_PARENT_FRAME checks and just define BaseFrame as
wxMDIParentFrame (instead of the default wxFrame) once when using this.

Also don't pass hardcoded position and size to the frame ctor, we don't
need to do this at all and it's actually harmful in high DPI as the
hardcoded size was too small.

No real changes, just simplify the code.