Revert "Fix wxDataViewCtrl border in dark mode under MSW"
This reverts commit 9e1b62779757a83acd75304e7c3defb40bd6a99a which fixed
wxDataViewCtrl border in dark mode (see #24624), but broke it in light
mode. It is possible to revert it now because the original bug was fixed
in a more general way in a later fbc76d6 (Don't use themed borders in
dark mode in wxMSW, 2024-09-25), so this doesn't break dark mode again.

Closes #25532.