Make empty wxAcceleratorTable always invalid
In addition to the changes from the previous commit, which made it
invalid if it was created with 0 entries, also make it invalid if it had
some entries before but the last of them was removed.

This seems more logical as it maintains the invariant that a non-empty
wxAcceleratorTable always has some entries.