Replace hex and octal escapes with literal UTF-8 characters
This makes the code more readable and there should be no reason not to
use UTF-8 in string literals any longer.

Also replace wxHAVE_U_ESCAPE (all still supported compilers have it)
with wxMUST_USE_U_ESCAPE, which is only set for MSVS 2015 currently and
will be removed when we drop support for it and can use literal Unicode
characters in the wide strings everywhere.