Don't crash on start up if macOS system font encoding is unknown
If the encoding is not recognized (which is the case at least for the
encoding used for Persian/Farsi), wxLocale::GetSystemEncoding() returns
wxFONTENCODING_DEFAULT and wxFont::SetDefaultEncoding() must not be
called in this case, as it asserts and, due to a separate bug, crashes
as the GUI is not fully initialized yet when wxApp::Initialize() is
executing.

See #25561.

(cherry picked from commit 5de590ffaf9ca1ebe8a619986f0217f3a44af8e5)