Fix overwriting IDs of buttons added to wxPGMultiButton
wxPGMultiButton::Add() replaced all negative IDs passed to it with
wxID_ANY, which was wrong as negative IDs can also be used and actually
are used for e.g. the controls created from XRC.

So avoid overwriting the IDs at all and just use wxID_ANY instead of
previously hardcoded "-2" as the default ID value in Add().

Closes #25514.