Fix regression in wxTopLevelWindow::Fit() when not using sizers
The changes of bcb9c19358 (Improve GetBestSize() for TLWs containing a
single child, 2022-12-14), fixing #22983, resulted in a regression in
Fit() behaviour in the same case, i.e. when a TLW contained a single
child and didn't use sizers.

Reimplement Fit() in this special case as well to make its behaviour
more logical and more compatible with the previous wx versions.

Closes #25537.