Fix disabled wxStaticText inside a dialog in wxMSW dark mode
The control wasn't rendered using the desired "disabled" colour because
wxControl::DoMSWControlColor() didn't take it into account in dark mode
unless m_hasFgCol was set and always used wxSYS_COLOUR_LISTBOXTEXT in
this case.

This didn't make much sense, as GetForegroundColour() should work
correctly even in dark mode, and made the behaviour of the code with
respect to the foreground colour inconsistent depending on whether
background brush was inherited from the parent (as was the case when the
parent was a wxFrame) or not (when it was a wxDialog).

So just use GetForegroundColour() in all cases, this is simpler and more
consistent and hopefully shouldn't break anything -- but if it does, it
would mean that GetForegroundColour() implementation needs fixing in
dark mode.

Closes #25574.