Also use control background in wxMSW dark mode
For the same reasons as were given for the foreground colour in the
previous commit, we should be using GetBackgroundColour() instead of
hard coding wxSYS_COLOUR_LISTBOX as background for the controls without
custom background brush in dark mode.

And, as with the parent commit, if this breaks something, it would
likely mean that GetBackgroundColour() needs fixing. But assuming it's
implemented correctly, this commit shouldn't change anything and just
simplify the code.