Don't call MSWGetCreateWindowFlags() unnecessarily
There is no need to call it only to pass to MSWCreateControl() overload
taking it when we can use another, simpler, overload which determines
the flags and extended flags to use itself -- and so calls MSWGetStyle()
only once instead of doing it twice as before, as it was still called
from MSWCreateControl() to fill in the extended style.

No real changes, this is just a micro-optimization.