Fix formatting of wxLongStringProperty documentation
It was completely broken because it's apparently impossible to format
"\\" in monospace font with Doxygen -- at least all my attempts to do
it, using either "@c" or "`...`" or "<tt>...</tt>" and amounts of
backslashes varying from 1 to 8 only resulted in either Doxygen warnings
and broken output or just broken output without warnings.

Just write that the backslash is doubled instead.