Make all MSWCreateControl() parameters mandatory
None of the existing callers of this function use default values for its
parameters and we don't want to make this function more convenient to
call as we want to discourage its use in favour of another overload.

Also update comments: the second overload is not really deprecated, but
the first one is still preferred.

No real changes.