Use "COMBOBOX" class explicitly for wxComboCtrl theme
The previous commit was not quite correct as we need to specify that we
want to use COMBOBOX explicitly in this control, unlike in the native
combobox which seems to use it implicitly internally.

In practice, this makes combobox-specific parts, such as CB_BORDER, work
correctly in dark mode instead of using some fallback, which resulted in
them being too white.