Remove wxComboCtrlBase::GetBackgroundColour()
It can't override the base class function because it is not virtual and
it's confusing and dangerous to have a function with the same name in
the derived class not overriding the base class one.

Instead, just call wxControl::SetBackgroundColour() when setting the
colour, then the default implementation of GetBackgroundColour() works
correctly and we don't need to do anything.