Fix possible confusion between bytes and characters in wxUILocale
The code used "valueSize", which is the size in bytes, as size of the
array of WCHAR, which is supposed to be expressed in elements, i.e.
(wide) characters in this case.

Make the intention clear by introducing an explicitly named "numChars".

This also means that we allocate enough space for 256, and not 128,
characters here, which was probably the intention.

We may still want to change this code to allocate the required size if
256 is not enough, even if this seems unlikely in practice.