Use wxRegKey to ensure that HKEY is always closed
No real changes, just simplify the code a bit by using wxRegKey instead
of opening and closing the key manually.

It would be even better to add support for REG_MULTI_SZ to wxRegKey
itself and use it directly here, but for now keep the code handling the
values of this type here as we don't need it anywhere else.

This commit is best viewed ignoring whitespace-only changes.