Fix multiple wxTreeCtrl::EnsureVisible() while frozen in wxMSW
Since the changes of e8712b3c56 (Make wxTreeCtrl::EnsureVisible() work
while frozen in wxMSW, 2019-07-09) calling EnsureVisible() more than
once while wxTreeCtrl was frozen didn't work any more, as we remembered
just the last item to make visible on thaw.

This was obviously wrong as it's perfectly possible for the application
to want to make multiple items visible while the control is frozen, e.g.
to highlight all items matching some criteria, so make it work again by
extending the fix to use a vector of items to make visible on thaw
instead of just a single one.

See #18435.