Revert "Remove wxGtkTreeModelNode::GetNodesCount()"
This reverts commit adbb1e5348a969f01d3f9d4d2efff85bb5e99b8e because its
assumption ("m_nodes and m_children must be always in sync") wasn't
actually true, "m_nodes" can be empty even if "m_children" is not if the
node hadn't been expanded yet.

Actually, checking m_nodes size is important as not doing it could
result in trying to access non-existent element of this container and
crashing.

Closes #25625.