Fix losing text of generic wxListCtrl items when copying them
During the switch from copy to move ctor of wxListCtrl in a5777cdcbd
(Make wxListItemData movable and not copyable, 2023-01-19) copying the
item text was simply lost, resulting in items forgetting their text when
they were moved internally by the vector in which they are stored, which
could result in e.g. their text becoming empty when a new column was
appended.

Fix this by adding the missing text (and also image and user data) field
to the move ctor and move assignment operator.

Closes #25519.