Restore compatibility with code using wxPG_XXX flags
Defining wxPG_XXX of type different from (and not convertible to) "int"
wasn't enough to make all code using these constants compile, as it
could rely on storing them in int variables.

Change the type of these compatibility constants to "int" to be really
compatible and also remove deprecation messages from them, as getting
such message when passing them to any function taking a new style enum
is enough, there is no need for multiple deprecation messages for the
same line of code.

Closes #25627.